refactor(utils): Move report related fns in report
