+++ packages/utils/src/index.js
@@ -3,4 +3,5 @@ export * from './config';
 export * from './jobs';
 export * from './lighthouse';
 export * from './metrics';
+export * from './report';
 export * from './utils';
