+++ packages/utils/src/metrics/index.js
@@ -1,3 +1,2 @@
 export * from './get-metric-changed';
 export * from './get-metric-type';
-export * from './merge-runs-by-id';
