+++ packages/utils/src/utils/index.js
@@ -1,4 +1,3 @@
-export * from './create-report';
 export * from './delta';
 export * from './date';
 export * from './format';
